+++ stagecraft/apps/datasets/admin/__init__.py
@@ -1,4 +1,3 @@
 from stagecraft.apps.datasets.admin.data_group import DataGroupAdmin
 from stagecraft.apps.datasets.admin.data_set import DataSetAdmin
 from stagecraft.apps.datasets.admin.data_type import DataTypeAdmin
-from stagecraft.apps.datasets.admin.backdrop_user import BackdropUserAdmin
