--- stagecraft/apps/datasets/tests/models/test_backdrop_user.py
@@ -1,35 +1,0 @@
-# encoding: utf-8
-# See https://docs.djangoproject.com/en/1.6/topics/testing/tools/
-
-from __future__ import unicode_literals
-
-from django.test import TestCase
-
-from django.core.exceptions import ValidationError
-
-from nose.tools import assert_raises, assert_equal
-
-from stagecraft.apps.datasets.models import(
-    BackdropUser, DataSet)
-
-
-class BackdropUserTestCase(TestCase):
-    fixtures = ['test_import_users_datasets.json']
-
-    def test_user_email_must_be_unique(self):
-        a = BackdropUser.objects.create(email='email@email.com')
-        a.validate_unique()
-
-        b = BackdropUser(email='email@email.com')
-        assert_raises(ValidationError, lambda: b.validate_unique())
-
-    def test_serialize_returns_serialized_user(self):
-        a = BackdropUser.objects.create(email='email@blah.net')
-        a.data_sets.add(DataSet.objects.get(name="evl_customer_satisfaction"))
-        a.data_sets.add(DataSet.objects.get(name="lpa_volumes"))
-        expected_response = {
-            'email': 'email@blah.net',
-            'data_sets': ['evl_customer_satisfaction', 'lpa_volumes']
-        }
-
-        assert_equal(a.serialize(), expected_response)
